+++ src/components/Pages/ClosedLoopMaf.vue
@@ -8,19 +8,66 @@ Remember to:
 -->
 
 <template>
-    <h1>Low Maf</h1>
+    <h1>{{title}}</h1>
     <img alt="Sedona Red Pearl Bugeye Impreza WRX" height="300" width="500" src="../../assets/carpic_0.jpg">
+    <br>
+    <br>
+    <CSV_Input @csvProcessed="buildObjects($event)" />
+    <br>
+    <br>
+    <CL_MAF_Output v-if="receivedData" :scales="scales" />
 </template>
 
 <script lang="ts">
 //Just regular TS here...
 
-import { Vue } from 'vue-class-component';
+import { Vue, Options } from 'vue-class-component';
+import CL_MAF_Output from './ClosedLoopMaf/CL_MAF_Output.vue'
+import CSV_Input from '../../components/CSV_Input.vue'
+
+import LowMafLog from './ClosedLoopMaf/lowMafLog';
+import LowMafOutput from './ClosedLoopMaf/lowMafOutput'
+
+@Options({
+  components: {
+    CL_MAF_Output,
+    CSV_Input
+  },
+})
 
 export default class ClosedLoopMaf extends Vue {
     title: string = 'Closed Loop Mass Air Flow Scaling';
+    receivedData: boolean = false;
+    scales: LowMafOutput[] = [];
+
+    buildObjects(data: {categories: string[], lines: string[]}){
+        this.receivedData = false;
+        let logs: LowMafLog[] = [];
+        data.lines.forEach(line => {
+            if (line && line != ""){
+                let lineData = line.split(',');
+                let log = new LowMafLog(data.categories, lineData);
+                logs.push(log);
+            }
+        })
+            
+        this.apiRequest(logs);
+    }
 
 
+    apiRequest(logs: LowMafLog[]){
+        const apiUrl = "http://localhost:8000/api/analyze/0/";
+        const requestOptions = {
+            method: "POST",
+            headers: { "Content-Type": "application/json" },
+            body: JSON.stringify(logs)
+        };
+        fetch(apiUrl, requestOptions).then(response => response.json())
+        .then(inf => {
+        this.scales = inf;
+        this.receivedData = true;
+        });
+    }
 }
 
 </script>
