--- src/components/Pages/ClosedLoopMaf/InputBox.vue
@@ -1,125 +1,0 @@
-<template>
-  <div class="hello">
-    <h1>Low-RPM MAF Scaling</h1>
-    <input type="file" v-on:change="uploadFile($event)">
-    <button v-if="csvSelected" v-on:click="normalize()">Normalize</button>
-    <br>
-    <br>
-    <table v-if="receivedData" class="table">
-      <thead>
-        <tr>
-          <th>Voltage</th>
-          <th>Scale Value</th>
-          <th>Data Points</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr v-for="scale of scales" v-bind:key="scale">
-          <td><b>{{scale.MafVoltage}}</b></td>
-          <td>{{scale.Correction}}</td>
-          <td>{{scale.Frequency}}</td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
-</template>
-
-<script lang="ts">
-import LowMafLog from '../lowMafLog';
-import LowMafOutput from '../lowMafOuput'
-import { Vue } from 'vue-class-component';
-
-export default class InputBox extends Vue{
-  public csvSelected: boolean = false;
-  public receivedData: boolean = false;
-  private file!: File;
-  public logs: LowMafLog[] = [];
-  public scales: LowMafOutput[] = [];
-
-  uploadFile(event: any) {
-    console.log(event.target);
-    console.log("File selected");
-    this.file = event.target.files[0];
-    if(this.file.type != 'text/csv'){
-      this.csvSelected = false;
-      console.log("Not a CSV");
-    }else{
-      this.csvSelected = true;
-      console.log("ayye good shit");
-      console.log("You uploaded: " + this.file.name);
-      console.log(this.file);
-    }
-  }
-
-  normalize() {
-    console.log("Checking Data for Errors and Formatting to JSON...");
-    this.readFile();
-  }
-
-  buildJson(data: string) {
-    //console.log(data);
-    let lines = data.split(/\r?\n/);
-    let categories = lines[0].split(',');
-    //console.log(categories);
-    let logs = [];
-    for(let i = 1; i < lines.length; i++){
-      if (lines[i] && lines[i] != ""){
-        let lineData = lines[i].split(',');
-        let log = new LowMafLog(categories, lineData);
-        logs.push(log);
-      }
-    }
-    console.log(JSON.stringify(logs));
-    this.apiRequest(logs);
-  }
-
-  apiRequest(logs: LowMafLog[]){
-    const apiUrl = "http://localhost:8000/api/analyze/0/";
-    const requestOptions = {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(logs)
-  };
-    fetch(apiUrl, requestOptions).then(response => response.json())
-    .then(inf => {
-      this.scales = inf;
-      this.receivedData = true;
-    });
-  }
-  
-  readFile() {
-    //console.log(this.file);
-    let fileReader = new FileReader(); 
-    fileReader.readAsText(this.file); 
-    fileReader.onloadend = (() => {
-      
-      if(typeof(fileReader.result) == 'string'){
-        this.buildJson(fileReader.result);
-      }
-    }); 
-
-    fileReader.onerror = function() {
-      alert(fileReader.error);
-    };
-  }
-
-}
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-h3 {
-  margin: 40px 0 0;
-}
-ul {
-  list-style-type: none;
-  padding: 0;
-}
-li {
-  display: inline-block;
-  margin: 0 10px;
-}
-a {
-  color: #42b983;
-}
-</style>
